requirements/base.txt
django-grappelli==2.12.2
djangorestframework==3.9.0
Django==2.2
django-rest-auth==0.9.3
django-model-utils==3.1.2
psycopg2==2.7.6.1
django-cors-headers==2.4.0
requests==2.21.0
beautifulsoup4==4.7.1
celery==4.2.1
redis==3.2.0
dj-database-url==0.5.0
django-reversion==3.0.3
Pillow==5.4.1
django-ckeditor==5.6.1
django-ipware==2.1.0
Willow==1.1
phonenumbers==8.10.8
raven==6.10.0
Unidecode==1.0.23
django-autocomplete-light==3.3.2
django-import-export==1.2.0
newrelic==4.20.0.120
googleads==20.0.0
authy==2.2.3
web3==5.11.0